Using YCM, plus TernJS.
This somewhat replaces SuperTab, but the autocomplete menu shows up
automatically. I believe there are options to prevent that and use a key
to trigger it, which I will be playing around with in the near future.

Note that YCM includes a Python completer, so something analagous to
Tern is not necessary.